Add dependency on fastscroller, use their scrollbar, remove platform-specific code
