+++ demo-common/build.gradle.kts
@@ -11,19 +11,8 @@ kotlin {
             dependencies {
                 implementation(project(":data-table-material3"))
                 implementation(compose.material3)
+                implementation(libs.fastscroller.m3)
             }
-        }
-
-        val skikoMain by creating {
-            dependsOn(commonMain)
-        }
-
-        val jvmMain by getting {
-            dependsOn(skikoMain)
-        }
-
-        val jsMain by getting {
-            dependsOn(skikoMain)
         }
     }
 }
