--- demo-common/src/androidMain/kotlin/com/seanproctor/datatable/demo/Scrollbar.android.kt
@@ -1,11 +1,0 @@
-package com.seanproctor.datatable.demo
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import com.seanproctor.datatable.DataTableScrollState
-
-@Composable
-actual fun VerticalScrollbar(scrollState: DataTableScrollState, modifier: Modifier) {}
-
-@Composable
-actual fun HorizontalScrollbar(scrollState: DataTableScrollState, modifier: Modifier) {}
