+++ gradle/libs.versions.toml
@@ -18,11 +18,12 @@ datatable-android-application = { id = "datatable.android.application" }
 
 [versions]
 
-activityCompose = "1.10.0"
+activityCompose = "1.10.1"
 androidGradlePlugin = "8.7.3"
 androidTools = "31.7.3"
 androidxLintGradle = "1.0.0-alpha03"
 compose = "1.7.3"
+fastscroller = "0.3.0"
 kotlin = "2.1.10"
 truth = "1.4.4"
 vanniktech = "0.30.0"
@@ -30,6 +31,7 @@ vanniktech = "0.30.0"
 [libraries]
 
 activity-compose = { module = "androidx.activity:activity-compose", version.ref = "activityCompose" }
+fastscroller-m3 = { module = "io.github.oikvpqya.compose.fastscroller:fastscroller-material3", version.ref = "fastscroller" }
 truth = { group = "com.google.truth", name = "truth", version.ref = "truth" }
 
 # Dependencies of the included build-logic
